// src/app/admin/pages/logs/log-name.enum.ts
export enum LogName {
  UserCreated = 'user.created',
  UserUpdated = 'user.updated',
  UserDeleted = 'user.deleted',
  StudentExtraDataCreated = 'studentExtraData.created',
  StudentExtraDataUpdated = 'studentExtraData.updated',
  PaperCreated = 'paper.created',
  PaperUpdated = 'paper.updated',
  PaperDeleted = 'paper.deleted',
  PaperSubmitted = 'paper.submitted',
  PaperUnsubmitted = 'paper.unsubmitted',
  PaperValidated = 'paper.validated',
  PaperInvalidated = 'paper.invalidated',
  PaperCancelledValidation = 'paper.cancelledValidation',
  PaperAssigned = 'paper.assigned',
  PaperUnassigned = 'paper.unassigned',
  PaperGraded = 'paper.graded',
  DocumentCreated = 'document.created',
  DocumentContentUpdated = 'document.contentUpdated',
  DocumentDeleted = 'document.deleted',
  DocumentReuploadRequestCreated = 'documentReuploadRequest.created',
  DocumentReuploadRequestDeleted = 'documentReuploadRequest.deleted',
}